fixed web import

+++ native/counter-ffi/src/lib.rs
@@ -1,5 +1,5 @@
 
 #[no_mangle]
-pub extern "C" fn increment(val: i64) -> i64 {
+pub extern "C" fn increment(val: i32) -> i32 {
     counter::increment(val)
 }
